+++ @xen-orchestra/lite/src/views/pool/PoolTasksView.vue
@@ -2,7 +2,7 @@
   <UiCard :color="hasError ? 'error' : undefined">
     <UiTitle class="title-with-counter" type="h4">
       {{ $t('tasks') }}
-      <UiCounter :value="pendingTasks.length" color="info" />
+      <UiCounter large :value="pendingTasks.length" color="info" />
     </UiTitle>
     <TasksTable :finished-tasks="finishedTasks" :pending-tasks="pendingTasks" />
     <UiCardSpinner v-if="!isReady" />
@@ -13,10 +13,10 @@
 import TasksTable from '@/components/tasks/TasksTable.vue'
 import UiCard from '@/components/ui/UiCard.vue'
 import UiCardSpinner from '@/components/ui/UiCardSpinner.vue'
-import UiCounter from '@/components/ui/UiCounter.vue'
 import UiTitle from '@/components/ui/UiTitle.vue'
-import { useTaskCollection } from '@/stores/xen-api/task.store'
 import { usePageTitleStore } from '@/stores/page-title.store'
+import { useTaskCollection } from '@/stores/xen-api/task.store'
+import UiCounter from '@core/components/UiCounter.vue'
 import { useI18n } from 'vue-i18n'
 
 const { pendingTasks, finishedTasks, isReady, hasError } = useTaskCollection()
@@ -33,11 +33,7 @@ titleStore.setCount(() => pendingTasks.value.length)
   display: flex;
   align-items: center;
   margin-bottom: 1rem;
-  gap: 0.5rem;
-
-  .ui-counter {
-    font-size: 1.4rem;
-  }
+  gap: 1rem;
 }
 
 .ui-card {